More API documentation and improvements.
