+++ ReactiveKit/SignalProtocol+Filtering.swift
@@ -27,6 +27,8 @@ import Foundation
 extension SignalProtocol {
 
     /// Emit an element only if `interval` time passes without emitting another element.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#debounceTime](https://rxmarbles.com/#debounceTime)
     public func debounce(interval: Double, queue: DispatchQueue = DispatchQueue(label: "reactive_kit.debounce")) -> Signal<Element, Error> {
         return Signal { observer in
             var timerSubscription: Disposable? = nil
@@ -56,18 +58,24 @@ extension SignalProtocol {
     }
 
     /// Emit first element and then all elements that are not equal to their predecessor(s).
-    public func distinct(areDistinct: @escaping (Element, Element) -> Bool) -> Signal<Element, Error> {
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#distinctUntilChanged](https://rxmarbles.com/#distinctUntilChanged)
+    public func distinctUntilChanged(_ areDistinct: @escaping (Element, Element) -> Bool) -> Signal<Element, Error> {
         return zipPrevious().compactMap { (prev, next) -> Element? in
             prev == nil || areDistinct(prev!, next) ? next : nil
         }
     }
 
-    /// Emit only element at given index if such element is produced.
+    /// Emit only the element at given index (if such element is produced).
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#elementAt](https://rxmarbles.com/#elementAt)
     public func element(at index: Int) -> Signal<Element, Error> {
         return take(first: index + 1).last()
     }
 
-    /// Emit only elements that pass `include` test.
+    /// Emit only elements that pass the `include` test.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#filter](https://rxmarbles.com/#filter)
     public func filter(_ isIncluded: @escaping (Element) -> Bool) -> Signal<Element, Error> {
         return Signal { observer in
             return self.observe { event in
@@ -83,8 +91,8 @@ extension SignalProtocol {
         }
     }
 
-    /// Filters the signal by executing `isIncluded` in each element and
-    /// propagates that element only if the returned signal fires `true`.
+    /// Filter the signal by executing `isIncluded` in each element and
+    /// propagate that element only if the returned signal emits `true`.
     public func filter(_ isIncluded: @escaping (Element) -> SafeSignal<Bool>) -> Signal<Element, Error> {
         return flatMapLatest { element -> Signal<Element, Error> in
             return isIncluded(element)
@@ -101,17 +109,21 @@ extension SignalProtocol {
         }
     }
 
-    /// Emit only the first element generated by the signal and then complete.
+    /// Emit only the first element of the signal and then complete.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#first](https://rxmarbles.com/#first)
     public func first() -> Signal<Element, Error> {
         return take(first: 1)
     }
 
     /// Ignore all elements (just propagate terminal events).
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#ignoreElements](https://rxmarbles.com/#ignoreElements)
     public func ignoreElements() -> Signal<Element, Error> {
         return filter { _ in false }
     }
 
-    /// Ignore all terminal events (just propagate next events).
+    /// Ignore all terminal events (just propagate next events). The signal will never complete or error out.
     public func ignoreTerminal() -> Signal<Element, Error> {
         return Signal { observer in
             return self.observe { event in
@@ -122,12 +134,14 @@ extension SignalProtocol {
         }
     }
 
-    /// Emit only last element generated by the signal and then complete.
+    /// Emit only the last element of the signal and then complete.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#last](https://rxmarbles.com/#last)
     public func last() -> Signal<Element, Error> {
         return take(last: 1)
     }
 
-    /// Supress events while last event generated on other signal is `false`.
+    /// Supress events while the last element on the other signal is `false`.
     public func pausable<O: SignalProtocol>(by other: O) -> Signal<Element, Error> where O.Element == Bool {
         return Signal { observer in
             var allowed: Bool = true
@@ -144,7 +158,9 @@ extension SignalProtocol {
         }
     }
 
-    /// Periodically sample the signal and emit latest element from each interval.
+    /// Periodically sample the signal and emit the latest element from each interval.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#sample](https://rxmarbles.com/#sample)
     public func sample(interval: Double, on queue: DispatchQueue = DispatchQueue(label: "reactive_kit.sample")) -> Signal<Element, Error> {
         return Signal { observer in
             let serialDisposable = SerialDisposable(otherDisposable: nil)
@@ -175,6 +191,8 @@ extension SignalProtocol {
     }
 
     /// Suppress first `count` elements generated by the signal.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#skip](https://rxmarbles.com/#skip)
     public func skip(first count: Int) -> Signal<Element, Error> {
         return Signal { observer in
             var count = count
@@ -194,6 +212,8 @@ extension SignalProtocol {
     }
 
     /// Suppress last `count` elements generated by the signal.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#skip](https://rxmarbles.com/#skip)
     public func skip(last count: Int) -> Signal<Element, Error> {
         guard count > 0 else { return self.toSignal() }
         return Signal { observer in
@@ -230,6 +250,8 @@ extension SignalProtocol {
     }
 
     /// Emit only first `count` elements of the signal and then complete.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#take](https://rxmarbles.com/#take)
     public func take(first count: Int) -> Signal<Element, Error> {
         guard count > 0 else { return .completed() }
         return Signal { observer in
@@ -252,6 +274,8 @@ extension SignalProtocol {
     }
 
     /// Emit only last `count` elements of the signal and then complete.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#takeLast](https://rxmarbles.com/#takeLast)
     public func take(last count: Int) -> Signal<Element, Error> {
         return Signal { observer in
             var values: [Element] = []
@@ -277,8 +301,28 @@ extension SignalProtocol {
         }
     }
 
-    /// Emit elements of the receiver until the given signal sends an event (of any kind)
-    /// and then completes the receiver (subsequent events on the receiver are ignored).
+    /// Emit elements until the given closure returns first `false`.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#takeWhile](https://rxmarbles.com/#takeWhile)
+    public func take(while shouldContinue: @escaping (Element) -> Bool) -> Signal<Element, Error> {
+        return Signal { observer in
+            return self.observe { event in
+                switch event {
+                case .next(let element):
+                    if shouldContinue(element) {
+                        observer.next(element)
+                    } else {
+                        observer.completed()
+                    }
+                default:
+                    observer.on(event)
+                }
+            }
+        }
+    }
+
+    /// Emit elements until the given signal sends an event (of any kind)
+    /// and then complete and dispose the signal.
     public func take<S: SignalProtocol>(until signal: S) -> Signal<Element, Error> {
         return Signal { observer in
             let disposable = CompositeDisposable()
@@ -300,6 +344,8 @@ extension SignalProtocol {
     }
 
     /// Throttle the signal to emit at most one element per given `seconds` interval.
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#throttle](https://rxmarbles.com/#throttle)
     public func throttle(seconds: Double) -> Signal<Element, Error> {
         return Signal { observer in
             var lastEventTime: DispatchTime?
@@ -322,7 +368,9 @@ extension SignalProtocol {
 extension SignalProtocol where Element: Equatable {
 
     /// Emit first element and then all elements that are not equal to their predecessor(s).
-    public func distinct() -> Signal<Element, Error> {
-        return distinct(areDistinct: !=)
+    ///
+    /// Check out interactive example: [https://rxmarbles.com/#distinctUntilChanged](https://rxmarbles.com/#distinctUntilChanged)
+    public func distinctUntilChanged() -> Signal<Element, Error> {
+        return distinctUntilChanged(!=)
     }
 }
